requirements.txt
certifi==2020.6.20
chardet==3.0.4
colour==0.1.5
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
idna==2.9
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.2
mpmath==1.1.0
networkx==2.4
netwulf==0.1.4
numpy==1.19.0
pandas==1.0.5
patsy==0.5.1
pipreqs==0.4.10
pyparsing==2.4.7
python-dateutil==2.8.1
python-ternary==1.0.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
simplejson==3.17.0
six==1.15.0
sklearn==0.0
statsmodels==0.11.1
sympy==1.6
threadpoolctl==2.1.0
urllib3==1.25.9
yarg==0.1.9
